+ Heroku prepa, new tests & - Exec time
Application is ready to be deployed to Heroku : ruby version added to
the Gemfile and Procfile created.
New tests for the limit functionnality have been implemented.
The execution timer has been deleted because its useless in production
mode.
The documentation has been modified and is now available in PDF format.